Added admin game/platform management.  Began admin user management.

+++ db/migrate/20090808154801_create_event_games.rb
@@ -5,7 +5,7 @@ class CreateEventGames < ActiveRecord::Migration
 	  t.integer :event_id
 	  t.integer :game_id
 	  t.integer :platform_id
-	  t.float :fee, :default => 0
+	  t.decimal :fee, :default => 0
 	  t.text :info
 	  t.datetime :start_at
 	  t.datetime :end_at
